+++ R/chart_zscore.R
@@ -11,7 +11,8 @@
 #' Frequency of seasonality "yearweek" (DEFAULT). "yearmonth", "yearquarter"
 #' @param output
 #' "stl" for STL decomposition chart,
-#' "stats" for STL statistical test results.
+#' "stats" for STL fitted statistics.
+#' "res" for STL fitted data.
 #' "zscore" for residuals Z-score,
 #' "seasonal" for standard seasonal chart.
 #' @param chart
@@ -27,6 +28,7 @@
 #' title <- "NGLower48"
 #' chart_zscore(df = df, title = " ",per = "yearweek", output = "stl", chart = "seasons")
 #' chart_zscore(df = df, title = " ",per = "yearweek", output = "stats", chart = "seasons")
+#' chart_zscore(df = df, title = " ",per = "yearweek", output = "res", chart = "seasons")
 #' chart_zscore(df = df, title = " ",per = "yearweek", output = "zscore", chart = "seasons")
 #' chart_zscore(df = df, title = " ",per = "yearweek", output = "seasonal", chart = "seasons")
 #' }
@@ -63,10 +65,21 @@ chart_zscore <- function(df = df, title = "NG Storage Z Score", per = "yearweek"
     if (per %in% c("yearweek")) {x <- x %>% tsibble::index_by(freq = ~yearweek(.))}
     if (per %in% c("yearmonth")) {x <- x %>% tsibble::index_by(freq = ~yearmonth(.))}
     if (per %in% c("yearquarter")) {x <- x %>% tsibble::index_by(freq = ~yearquarter(.))}
-    x <- x %>%  dplyr::summarise(value = mean(value)) %>%
+    x <- x %>% dplyr::summarise(value = mean(value)) %>%
+      fabletools::features(value,feasts::feat_stl)
+    return(x)
+  }
+
+  if (output == "res") {
+    x <- rbind(df,df %>% dplyr::mutate(series = title)) %>%
+      tsibble::as_tsibble(key=series, index = date) %>%
+      tsibble::group_by_key() #%>%
+    if (per %in% c("yearweek")) {x <- x %>% tsibble::index_by(freq = ~yearweek(.))}
+    if (per %in% c("yearmonth")) {x <- x %>% tsibble::index_by(freq = ~yearmonth(.))}
+    if (per %in% c("yearquarter")) {x <- x %>% tsibble::index_by(freq = ~yearquarter(.))}
+    x <- x %>% dplyr::summarise(value = mean(value)) %>%
       fabletools::model(feasts::STL(value ~ season(window = Inf))) %>%
-      fabletools::components() %>%
-      dplyr::slice(1)
+      fabletools::components()
     return(x)
   }